apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "dependency-injector>=4.48.2",
    "fastapi[standard]>=0.117.1",
    "logger",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "sqladmin>=0.21.0",
    "sqlalchemy>=2.0.43",
    "testcontainers[postgresql]>=4.13.2",
]

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
    "pytest>=8.4.2",
    "ruff>=0.13.2",
    "typer>=0.19.2",
]


[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403", "F405"]



[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.12"
warn_return_any = true

[tool.uv.sources]
logger = { path = "../../packages/logger", editable = true }

[tool.pytest.ini_options]
python_files = ["test*.py", "tests.py"]
